Add progress reporting and progressive input file reading
This should probably be a patch series, but it's simple and small enough that
it wasn't worth splitting it up.

The issues referred to in the following description are:

    Inefficient upload of large files
    http://code.google.com/p/python-scribd/issues/detail?id=4

    No upload progress indication
    http://code.google.com/p/python-scribd/issues/detail?id=5

Fix #4 part 1: Accept a file-like object passed by the user as a buffer to use
instead of a string. If no buffer is passed, a StringIO is used to match the
old behaviour. I'd prefer to use a tempfile, but didn't want to change existing
behaviour. The upload.py example shows how a tempfile is used.

Fix #4 part 2: Instead of reading the whole input file in the upload() routine,
the file object is passed through to multipart.encode_multipart_formdata, which
now handles both string and file-like payloads. The file is copied to the
output buffer in blocks, so it's never all in memory.

Fix required for both #4 and #5: Send the data in blocks rather than using
httplib.request(...) to send the whole request (headers and data) in one go.
Request is sent with h.putrequest(), then headers with h.putheader() and
h.endheaders(), then finally data is sent in blocks by looping over h.send().
This means we don't have to have the whole request in memory at once to pass to
h.request()

Fix #5: Accept a `progress_callback` parameter. If passed, call the callback
for for every loop over h.send() to indicate progress. The callback takes two
arguments, as documented; one the number of bytes sent so far and one the total
number of bytes to send. Document this callback in the upload() method; it's
actually accepted for all methods, but upload() is the only one you're likely
to want to use it for. Update upload.py to show how to write a simple progress
function.